Completed Chapter 6 (Stats)
Eventually will have to retune how stats naturally drain over time. Currently, it decreases the stat by a percentage of the max, which means increasing the total max currently doesn't improve anything for the player. Change this to decrease a static amount or add different forms of stat damage later that drain a static amount to give meaning to a higher maximum.